背包接口数据会整合在线玩家数据

+++ gms-server/src/main/java/org/gms/model/dto/InventoryEquipRtnDTO.java
@@ -21,71 +21,71 @@ public class InventoryEquipRtnDTO {
     /**
      * 砸券次数，对应upgradeslots
      */
-    private Integer upgradeSlots;
+    private Byte upgradeSlots;
     /**
      * 装备等级，对应level
      */
-    private Integer level;
+    private Byte level;
     /**
      * 力量，对应str
      */
-    private Integer attStr;
+    private Short attStr;
     /**
      * 敏捷，对应dex
      */
-    private Integer attDex;
+    private Short attDex;
     /**
      * 智力，对应int
      */
-    private Integer attInt;
+    private Short attInt;
     /**
      * 运气，对应luk
      */
-    private Integer attLuk;
+    private Short attLuk;
     /**
      * 血量，对应hp
      */
-    private Integer hp;
+    private Short hp;
     /**
      * 蓝量，对应mp
      */
-    private Integer mp;
+    private Short mp;
     /**
      * 物理攻击，对应watk
      */
-    private Integer pAtk;
+    private Short pAtk;
     /**
      * 魔法攻击，对应matk
      */
-    private Integer mAtk;
+    private Short mAtk;
     /**
      * 物理防御，对应wdef
      */
-    private Integer pDef;
+    private Short pDef;
     /**
      * 魔法防御，对应mDef
      */
-    private Integer mdef;
+    private Short mdef;
     /**
      * 命中，对应acc
      */
-    private Integer acc;
+    private Short acc;
     /**
      * 回避，对应avoid
      */
-    private Integer avoid;
+    private Short avoid;
     /**
      * 攻速，对应hands
      */
-    private Integer hands;
+    private Short hands;
     /**
      * 移速，对应speed
      */
-    private Integer speed;
+    private Short speed;
     /**
      * 跳跃力，对应jump
      */
-    private Integer jump;
+    private Short jump;
     /**
      * 锁定，对应locked
      */
@@ -93,15 +93,15 @@ public class InventoryEquipRtnDTO {
     /**
      * 锤子次数，对应vicious
      */
-    private Long vicious;
+    private Short vicious;
     /**
      * 装备升级等级，对应itemlevel
      */
-    private Integer itemLevel;
+    private Byte itemLevel;
     /**
      * 装备升级经验，对应itemexp
      */
-    private Long itemExp;
+    private Integer itemExp;
     /**
      * 戒指id，对应ringid
      */
